Impmement the %allow-range mechanism internally to use int64_t type variables for the lower and upper range value.
With the newer implementation the %allow-range values in rtrmgr template
can be any range up to [0..4294967295] for u32 or
[-2147483648..2147483647] for i32.

Bug found by:	Bruce M Simpson, Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7761 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/rtrmgr/template_base_command.hh
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-// $XORP: xorp/rtrmgr/template_base_command.hh,v 1.6 2005/11/17 08:31:57 pavlin Exp $
+// $XORP: xorp/rtrmgr/template_base_command.hh,v 1.7 2005/11/27 05:43:36 pavlin Exp $
 
 #ifndef __RTRMGR_TEMPLATE_BASE_COMMAND_HH__
 #define __RTRMGR_TEMPLATE_BASE_COMMAND_HH__
@@ -133,7 +133,7 @@ public:
 
 private:
     // Map between the pair of [lower, upper] boundaries and the help string
-    typedef map<pair<int32_t, int32_t>, string> Filter;
+    typedef map<pair<int64_t, int64_t>, string> Filter;
     map<string, Filter> _filters;	// Map between a varname and a filter
 };
